gan-api/requirements.txt

fastapi>=0.68.0
uvicorn>=0.15.0
torch>=1.12.0
pillow>=9.0.0
numpy>=1.22.0
python-multipart
slowapi==0.1.7
limits==2.7.0